pyproject.toml
[project]
name = "sae"
version = "0.1.0"
description = "A2A-native legal agent for contract clause review and risk analysis"
readme = "README.md"
authors = [
    { name = "Egor Vert", email = "theegorvert@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.12",

    # Agent framework
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",

    # Vector database
    "pinecone-client>=5.0.0",

    # Data validation & settings
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # HTTP client
    "httpx>=0.27.0",

    # Document parsing
    "python-docx>=1.1.0",
    "pypdf>=5.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",

    # Rate limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sae"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.coverage.run]
source = ["src/sae"]
branch = true
omit = [
    "*/__init__.py",
    "src/sae/py.typed",
]

[tool.coverage.report]
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true
